Refactor project design, remove duplicate code, implement temporary package structure
