+++ app/src/main/java/mozilla/org/webmaker/WebmakerApplication.java
@@ -3,12 +3,16 @@ package mozilla.org.webmaker;
 import android.app.Application;
 import android.util.Log;
 
+import mozilla.org.webmaker.activity.Editor;
+import mozilla.org.webmaker.activity.Map;
+import mozilla.org.webmaker.activity.Project;
+import mozilla.org.webmaker.activity.Tinker;
 import mozilla.org.webmaker.router.Router;
 
 public class WebmakerApplication extends Application {
 
     // Singleton
-    private static WebmakerApplication singleton;
+    private WebmakerApplication singleton;
     public WebmakerApplication getInstance(){
         return singleton;
     }
