+++ app/src/main/java/mozilla/org/webmaker/activity/Tinker.java
@@ -1,4 +1,7 @@
-package mozilla.org.webmaker;
+package mozilla.org.webmaker.activity;
+
+import mozilla.org.webmaker.R;
+import mozilla.org.webmaker.WebmakerActivity;
 
 public class Tinker extends WebmakerActivity {
     public Tinker() {
