+++ app/src/main/java/mozilla/org/webmaker/fragment/WebviewFragment.java
@@ -2,22 +2,14 @@ package mozilla.org.webmaker.fragment;
 
 import android.annotation.SuppressLint;
 import android.app.Fragment;
-import android.content.Context;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.animation.AlphaAnimation;
-import android.view.animation.Animation;
-import android.view.animation.DecelerateInterpolator;
-import android.webkit.JavascriptInterface;
-import android.webkit.WebView;
-import android.webkit.WebViewClient;
 import android.widget.RelativeLayout;
 
 import mozilla.org.webmaker.R;
-import mozilla.org.webmaker.WmWebView;
+import mozilla.org.webmaker.view.WebmakerWebView;
 
 @SuppressLint("SetJavaScriptEnabled")
 public class WebviewFragment extends Fragment {
@@ -26,7 +18,7 @@ public class WebviewFragment extends Fragment {
      * The fragment argument representing the section number for this fragment.
      */
     private static final String ARG_SECTION_NUMBER = "section_number";
-    private WmWebView mWebView = null;
+    private WebmakerWebView mWebView = null;
 
     /**
      * Returns a new instance of this fragment for the given section number.
@@ -39,25 +31,12 @@ public class WebviewFragment extends Fragment {
         return fragment;
     }
 
-    public class WebAppInterface {
-        Context mContext;
-
-        WebAppInterface(Context c) {
-            mContext = c;
-        }
-
-        @JavascriptInterface
-        public void logText(String txt){
-            Log.v("wm", txt);
-        }
-    }
-
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         View mView = inflater.inflate(R.layout.fragment_main, container, false);
         String sectionId = Integer.toString(super.getArguments().getInt(ARG_SECTION_NUMBER));
 
-        mWebView = new WmWebView(mView.getContext(), "section-" + sectionId);
+        mWebView = new WebmakerWebView(mView.getContext(), "section-" + sectionId);
         RelativeLayout layout = (RelativeLayout)mView.findViewById(R.id.webview_fragment);
         layout.addView(mWebView);
         return mView;
@@ -65,34 +44,9 @@ public class WebviewFragment extends Fragment {
 
     @Override
     public void onDestroyView() {
-        if (mWebView != null) {
-            mWebView.destroy();
-            mWebView = null;
-        }
+        assert mWebView != null;
+        mWebView.destroy();
+        mWebView = null;
         super.onDestroyView();
-    }
-
-    private void animate(final WebView view) {
-        Animation fadeIn = new AlphaAnimation(0, 1);
-        fadeIn.setInterpolator(new DecelerateInterpolator()); //add this
-        fadeIn.setDuration(1000);
-        view.startAnimation(fadeIn);
-    }
-
-    private class WebClient extends WebViewClient {
-
-        @Override
-        public boolean shouldOverrideUrlLoading(WebView view, String url) {
-            view.setVisibility(View.GONE);
-            return false;
-        }
-
-        @Override
-        public void onPageFinished(WebView view, String url) {
-            animate(view);
-            view.setVisibility(View.VISIBLE);
-            super.onPageFinished(view, url);
-        }
-
     }
 }
