+++ app/src/main/res/layout/editor_layout.xml
@@ -1,6 +1,6 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="mozilla.org.webmaker.Editor"
+    tools:context="mozilla.org.webmaker.activity.Editor"
     android:id="@+id/editor_layout">
 </RelativeLayout>
