+++ app/src/main/res/menu/menu_editor.xml
@@ -1,5 +1,5 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools" tools:context="mozilla.org.webmaker.Editor">
+    xmlns:tools="http://schemas.android.com/tools" tools:context="mozilla.org.webmaker.activity.Editor">
     <item android:id="@+id/action_settings" android:title="@string/action_settings"
         android:orderInCategory="100" android:showAsAction="never" />
 </menu>
