+++ app/src/main/res/menu/menu_project.xml
@@ -1,5 +1,5 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools" tools:context="mozilla.org.webmaker.Project">
+    xmlns:tools="http://schemas.android.com/tools" tools:context="mozilla.org.webmaker.activity.Project">
     <item android:id="@+id/action_settings" android:title="@string/action_settings"
         android:orderInCategory="100" android:showAsAction="never" />
 </menu>
